# 0-back/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9001",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/decisions-filtrage-qpc/search": {
      "post": {
        "tags": [
          "Décisions de filtrage QPC"
        ],
        "summary": "Search paginated décisions de filtrage QPC",
        "operationId": "searchDecisionFiltrage",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [
                "id.desc"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionFiltrageQpcSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDTODecisionFiltrageQpcDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "operationId": "registerUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/signin": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Authenticate a user",
        "operationId": "authenticateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Demander la réinitialisation du mot de passe",
        "operationId": "requestPasswordReset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/reset-password/confirm": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Confirmer la réinitialisation du mot de passe",
        "operationId": "confirmPasswordReset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/decisions-filtrage-qpc/{id}": {
      "get": {
        "tags": [
          "Décisions de filtrage QPC"
        ],
        "summary": "Get a décision de filtrage QPC by ID",
        "operationId": "getDecisionFiltrageById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionFiltrageQpcDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nom": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "prenom": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "tel": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ROLE_USER",
              "ROLE_MANAGER",
              "ROLE_ADMIN",
              "ROLE_SUPERVISEUR"
            ]
          }
        }
      },
...
      "DroitLiberteDTO": {
        "required": [
          "id",
          "texte"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "texte": {
            "type": "string"
          }
        }
      },
...
      "DecisionFiltrageQpcSearchRequest": {
        "type": "object",
        "properties": {
          "ordreJuridictionnel": {
            "type": "string",
            "enum": [
              "JUDICIAIRE",
              "ADMINISTRATIF"
            ]
          },
          "juridiction": {
            "type": "string",
            "enum": [
              "CONSEIL_ETAT",
              "COUR_DE_CASSATION"
            ]
          },
          "niveauFiltrage": {
            "type": "string",
            "enum": [
              "NONE",
              "PREMIER_ET_DERNIER",
              "DEUXIEME_FILTRAGE"
            ]
          },
          "chambreSousSectionId": {
            "type": "integer",
            "format": "int64"
          },
          "numeroChambresReuniesId": {
            "type": "integer",
            "format": "int64"
          },
          "niveauCompetenceId": {
            "type": "integer",
            "format": "int64"
          },
          "matiereId": {
            "type": "integer",
            "format": "int64"
          },
          "qualiteDemandeurId": {
            "type": "integer",
            "format": "int64"
          },
          "qualitePreciseDemandeurId": {
            "type": "integer",
            "format": "int64"
          },
          "decisionRenvoiId": {
            "type": "integer",
            "format": "int64"
          },
          "decisionNonRenvoiId": {
            "type": "integer",
            "format": "int64"
          },
          "applicationTheorieChangementCirconstancesId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DecisionFiltrageQpcDTO": {
        "required": [
          "applicationTheorieChangementCirconstances",
          "autresRemarques",
          "chambreSousSection",
          "createdAt",
          "dateFiltrage",
          "decisionNonRenvoi",
          "decisionRenvoi",
          "droitsLibertes",
          "formationJugement",
          "id",
          "identiteDemandeur",
          "juridiction",
          "loiOrigineDisposition",
          "matiere",
          "motsCles",
          "niveauCompetence",
          "niveauFiltrage",
          "nombreDroitsNonMentionnes",
          "numeroChambresReunies",
          "numeroDecision",
          "ordreJuridictionnel",
          "origineJuridictionnelleQpc",
          "qualiteDemandeur",
          "qualitePreciseDemandeur",
          "reference",
          "referenceDispositionsContestees",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "ordreJuridictionnel": {
            "type": "string",
            "enum": [
              "JUDICIAIRE",
              "ADMINISTRATIF"
            ]
          },
          "juridiction": {
            "type": "string",
            "enum": [
              "CONSEIL_ETAT",
              "COUR_DE_CASSATION"
            ]
          },
          "chambreSousSection": {
            "type": "string"
          },
          "numeroChambresReunies": {
            "type": "string"
          },
          "niveauFiltrage": {
            "type": "string",
            "enum": [
              "NONE",
              "PREMIER_ET_DERNIER",
              "DEUXIEME_FILTRAGE"
            ]
          },
          "origineJuridictionnelleQpc": {
            "type": "string"
          },
          "niveauCompetence": {
            "type": "string"
          },
          "dateFiltrage": {
            "type": "string",
            "format": "date"
          },
          "formationJugement": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "numeroDecision": {
            "type": "string"
          },
          "referenceDispositionsContestees": {
            "type": "string"
          },
          "loiOrigineDisposition": {
            "type": "string"
          },
          "matiere": {
            "type": "string"
          },
          "qualiteDemandeur": {
            "type": "string"
          },
          "qualitePreciseDemandeur": {
            "type": "string"
          },
          "identiteDemandeur": {
            "type": "string"
          },
          "decisionRenvoi": {
            "type": "string"
          },
          "decisionNonRenvoi": {
            "type": "string"
          },
          "applicationTheorieChangementCirconstances": {
            "type": "string"
          },
          "nombreDroitsNonMentionnes": {
            "type": "string"
          },
          "autresRemarques": {
            "type": "string"
          },
          "motsCles": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "droitsLibertes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DroitLiberteDTO"
            }
          }
        }
      },
      "PageDTODecisionFiltrageQpcDTO": {
        "required": [
          "content",
          "page",
          "size",
          "totalElements"
        ],
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DecisionFiltrageQpcDTO"
            }
          }
        }
      },
      "SignupRequest": {
        "required": [
          "email",
          "nom",
          "password",
          "prenom",
          "tel"
        ],
        "type": "object",
        "properties": {
          "prenom": {
            "type": "string"
          },
          "nom": {
            "type": "string"
          },
          "email": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "password": {
            "maxLength": 40,
            "minLength": 6,
            "type": "string"
          },
          "tel": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ResetPasswordRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "ResetPasswordConfirmRequest": {
        "required": [
          "hash",
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "hash": {
            "type": "string"
          },
          "newPassword": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          }
        }
      }
    }
  }
}